[FEAT] jestream apis out of preview [FIX] jetstream materialized views marked as beta APIs [FIX] `deliver_subject` was not part of ConsumerUpdateConfig [DOC] added additional hints about `bind` in the jestream doc
